fix: refactor type tests to be stripped
