+++ packages/model/src/-private/attr.ts
@@ -4,8 +4,6 @@
 import { assert } from '@ember/debug';
 import { computed } from '@ember/object';
 
-import { expectTypeOf } from 'expect-type';
-
 import { DEBUG } from '@ember-data/env';
 import { recordIdentifierFor } from '@ember-data/store';
 import { peekCache } from '@ember-data/store/-private';
@@ -40,7 +38,7 @@ import { isElementDescriptor } from './util';
  * @class NOTATHING
  * @typedoc
  */
-type AttrOptions<DV = PrimitiveValue | object | unknown[]> = {
+export type AttrOptions<DV = PrimitiveValue | object | unknown[]> = {
   /**
    * The default value for this attribute.
    *
@@ -149,9 +147,9 @@ type LooseTransformInstance<V, Raw, Name extends string> = {
 
   [TransformName]: Name;
 };
-type TransformHasType = { [TransformName]: string };
+export type TransformHasType = { [TransformName]: string };
 
-type TypedTransformInstance<V, T extends string> =
+export type TypedTransformInstance<V, T extends string> =
   | LooseTransformInstance<V, string, T>
   | LooseTransformInstance<V, number, T>
   | LooseTransformInstance<V, boolean, T>
@@ -161,12 +159,12 @@ type TypedTransformInstance<V, T extends string> =
 
 // see note on Explicit ANY above
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-type GetMaybeDeserializeValue<T> = T extends { deserialize: (...args: any[]) => unknown }
+export type GetMaybeDeserializeValue<T> = T extends { deserialize: (...args: any[]) => unknown }
   ? ReturnType<T['deserialize']>
   : never;
 
-type TypeFromInstance<T> = T extends TransformHasType ? T[typeof TransformName] : never;
-type OptionsFromInstance<T> = TypeFromInstance<T> extends never
+export type TypeFromInstance<T> = T extends TransformHasType ? T[typeof TransformName] : never;
+export type OptionsFromInstance<T> = TypeFromInstance<T> extends never
   ? never
   : GetMaybeDeserializeValue<T> extends never
     ? never
@@ -181,7 +179,7 @@ type OptionsFromInstance<T> = TypeFromInstance<T> extends never
  *
  * @typedoc
  */
-type DataDecorator = (target: object, key: string, desc?: DecoratorPropertyDescriptor) => void;
+export type DataDecorator = (target: object, key: string, desc?: DecoratorPropertyDescriptor) => void;
 
 /**
   `attr` defines an attribute on a [Model](/ember-data/release/classes/Model).
@@ -289,110 +287,3 @@ export function attr(
   // see note on DataDecorator for why void
   return isElementDescriptor(args) ? (_attr()(...args) as void) : _attr(type, options as object);
 }
-
-// ------------------------------
-//              💚
-// ==============================
-//          Type Tests
-// ==============================
-//              🐹
-// ⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇⬇
-
-expectTypeOf<{ defaultValue: () => object }>().toMatchTypeOf<AttrOptions<object>>();
-expectTypeOf<{ defaultValue: () => object }>().toMatchTypeOf<AttrOptions>();
-expectTypeOf<{ defaultValue: () => object }>().not.toMatchTypeOf<AttrOptions<object[]>>();
-expectTypeOf<{ defaultValue: () => object }>().not.toMatchTypeOf<AttrOptions<string>>();
-expectTypeOf<{ defaultValue: () => string }>().not.toMatchTypeOf<AttrOptions<object>>();
-
-type ExampleDateTransform = {
-  serialize(value: Date, options: { dateOnly: boolean }): string;
-  deserialize(value: string, options: { stripTimeZone?: boolean }): Date;
-  [TransformName]: 'date';
-};
-
-expectTypeOf<
-  TypedTransformInstance<GetMaybeDeserializeValue<ExampleDateTransform>, TypeFromInstance<ExampleDateTransform>>
->().toMatchTypeOf<TypedTransformInstance<Date, 'date'>>();
-expectTypeOf<ExampleDateTransform>().toMatchTypeOf<TypedTransformInstance<Date, 'date'>>();
-expectTypeOf<TypeFromInstance<ExampleDateTransform>>().toEqualTypeOf<'date'>();
-expectTypeOf<GetMaybeDeserializeValue<ExampleDateTransform>>().toEqualTypeOf<Date>();
-expectTypeOf<OptionsFromInstance<ExampleDateTransform>>().toMatchTypeOf<{
-  dateOnly: boolean;
-  stripTimeZone?: boolean;
-  defaultValue?: () => Date;
-}>();
-expectTypeOf({ dateOnly: true }).toMatchTypeOf<OptionsFromInstance<ExampleDateTransform>>();
-expectTypeOf({ dateOnly: true, stripTimeZone: true }).toMatchTypeOf<OptionsFromInstance<ExampleDateTransform>>();
-expectTypeOf({ dateOnly: true, stripTimeZone: true, defaultValue: () => new Date() }).toMatchTypeOf<
-  OptionsFromInstance<ExampleDateTransform>
->();
-
-expectTypeOf(attr({}, 'key', {})).toEqualTypeOf<void>();
-expectTypeOf(attr('string')).toEqualTypeOf<DataDecorator>();
-expectTypeOf(attr({})).toEqualTypeOf<DataDecorator>();
-expectTypeOf(attr()).toEqualTypeOf<DataDecorator>();
-
-expectTypeOf(attr('string', { defaultValue: 'hello' })).toEqualTypeOf<DataDecorator>();
-expectTypeOf(attr({ defaultValue: 'hello' })).toEqualTypeOf<DataDecorator>();
-expectTypeOf(attr('string', { defaultValue: () => ({}) })).toEqualTypeOf<DataDecorator>();
-expectTypeOf(attr({ defaultValue: () => ({}) })).toEqualTypeOf<DataDecorator>();
-
-expectTypeOf<TypeFromInstance<ExampleDateTransform>>().toEqualTypeOf<'date'>();
-expectTypeOf(attr<ExampleDateTransform>('date')).toEqualTypeOf<DataDecorator>();
-/* prettier-ignore */
-expectTypeOf(attr<ExampleDateTransform>('date', {
-  dateOnly: true,
-  // @ts-expect-error - defaultValue needs to be a Date so it can be serialized, so must be a function that produces one
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-  defaultValue: string
-})).toBeNever;
-
-/* prettier-ignore */
-expectTypeOf(
-  // @ts-expect-error
-  attr(
-    'string',
-    { defaultValue: ({}) }
-  )
-).toEqualTypeOf<DataDecorator>();
-
-/* prettier-ignore */
-expectTypeOf(attr<ExampleDateTransform>('date', {
-  // @ts-expect-error - no stateful default values
-  defaultValue: new Date()
-})).toBeNever;
-expectTypeOf(
-  attr<ExampleDateTransform>('date', { dateOnly: false, defaultValue: () => new Date() })
-).toEqualTypeOf<DataDecorator>();
-
-/* prettier-ignore */
-expectTypeOf(
-  attr(
-    // @ts-expect-error
-    { defaultValue: {} }
-  )
-).toBeNever;
-expectTypeOf(
-  attr(
-    // @ts-expect-error
-    1,
-    { defaultValue: 'hello' }
-  )
-).toBeNever;
-
-expectTypeOf(
-  (function () {
-    class User {
-      @attr() declare name: string;
-    }
-    return User;
-  })()
-).toMatchTypeOf<{ name: string }>();
-expectTypeOf(
-  (function () {
-    class User {
-      @attr declare name: string;
-    }
-    return User;
-  })()
-).toMatchTypeOf<{ name: string }>();
